% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change_SSfiles.R
\name{Change.SSfiles}
\alias{Change.SSfiles}
\title{Change SS files}
\usage{
Change.SSfiles(main.dir, sppname.in, file.make = c(1, 1, 1, 1, 1),
  starter.par = "N")
}
\arguments{
\item{sppname.in}{species name}

\item{starter.par}{switch to use the par file}
}
\description{
Change SS files
}
\author{
Jason Cope
}
