\name{change_type}
\alias{change_type}
\title{Sets the data type of a column to correspond to parsed sss metadata...}
\usage{change_type(sss, df, i)}
\description{Sets the data type of a column to correspond to parsed sss metadata}
\seealso{llply_colwise}
\keyword{internal}
\arguments{\item{sss}{Parsed .sss metadata information}
\item{df}{Parsed .asc data}
\item{i}{The column number of sss that should be processed}}
