\name{change_values}
\alias{change_values}
\title{Applies coded values to asc data, as described in sss metadata...}
\usage{change_values(sss, df)}
\description{Applies coded values to asc data, as described in sss metadata}
\keyword{internal}
\seealso{llply_colwise}
\arguments{\item{sss}{Parsed .sss metadata information}
\item{df}{Parsed .asc data}}
