\name{get_sss_record}
\alias{get_sss_record}
\title{Reads all "variables" inside the triple-s "record"...}
\usage{get_sss_record(xmlNode)}
\description{Reads all "variables" inside the triple-s "record"}
\details{This function parses the record node, extracts all variables
and creates a data frame with information about size, position, type, etc.}
\keyword{internal}
\arguments{\item{xmlNode}{XML node}}
