\name{get_variable_position}
\alias{get_variable_position}
\title{Reads metadata to find position of variables...}
\usage{get_variable_position(sss, asc, variable_row)}
\description{Reads metadata to find position of variables}
\arguments{\item{sss}{Parsed .sss metadata information}
\item{asc}{Parsed .asc data}
\item{variable_row}{Variable row}}
