% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-xml.R
\name{getSSScodes}
\alias{getSSScodes}
\title{Reads all "codes" inside the triple-s "record".}
\usage{
getSSScodes(x)
}
\arguments{
\item{x}{XML node}
}
\description{
This function parses the record node and extracts all "codes" nodes into a data.frame
}
\keyword{internal}

