\name{getSSSrecord}
\alias{getSSSrecord}
\title{Reads all "variables" inside the triple-s "record".}
\usage{
  getSSSrecord(xmlNode)
}
\arguments{
  \item{xmlNode}{XML node}
}
\description{
  This function parses the record node, extracts all
  variables and creates a data frame with information about
  size, position, type, etc.
}
\keyword{internal}

