% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess-asc-filename.R
\name{guess_asc_filename}
\alias{guess_asc_filename}
\title{Guess the data filename given the sss filename.}
\usage{
guess_asc_filename(x)
}
\arguments{
\item{x}{filename}
}
\description{
Guess the data filename given the sss filename.
}
