\name{llply_colwise}
\alias{llply_colwise}
\title{Applies a custom function to each column of a data...}
\usage{llply_colwise(sss, df, custom_function)}
\description{Applies a custom function to each column of a data.frame}
\keyword{internal}
\arguments{\item{sss}{Parsed .sss metadata information}
\item{df}{Parsed .asc data}
\item{custom_function}{A function that operates on a single column}}
\examples{# None}
