% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-xml.R
\name{namesMultiple}
\alias{namesMultiple}
\title{Reads all "codes" inside the triple-s "record"}
\usage{
namesMultiple(names, length, sep = "_")
}
\arguments{
\item{names}{character vector defining the string that separates field and subfield names}

\item{length}{Numeric vector that indicates the number of times each row must be repeated}

\item{sep}{Character vector defining the string that separates field and subfield names}
}
\description{
This function parses the record node and extracts all "codes" nodes into a data.frame
}
\keyword{internal}
