% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_s_prof.R
\name{opt.s.prof}
\alias{opt.s.prof}
\title{Solve for the SD to match rbeta to either lognormal or truncated normal}
\usage{
opt.s.prof(mean.val, mean.match, sd.match, seedit, logn = T)
}
\description{
Solve for the SD to match rbeta to either lognormal or truncated normal
}
\author{
Jason Cope
}
