\name{parse_multiple}
\alias{parse_multiple}
\title{Parse field of type "multiple"...}
\usage{parse_multiple(sss, df, i)}
\description{Parse field of type "multiple"}
\seealso{llply_colwise}
\arguments{\item{sss}{Parsed .sss metadata information}
\item{df}{Parsed .asc data}
\item{i}{The column number of sss that should be processed}}
