% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_solve.R
\name{r.beta.solve}
\alias{r.beta.solve}
\title{Objective function used to match SD from beta to either lognormal or truncated normal}
\usage{
r.beta.solve(sd.init, mean.init, mean.match, sd.match, seedit, logn = T)
}
\description{
Objective function used to match SD from beta to either lognormal or truncated normal
}
\author{
Jason Cope
}
