% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.R
\name{rbeta.ab}
\alias{rbeta.ab}
\title{Beta function}
\usage{
rbeta.ab(n, m, s, a, b)
}
\arguments{
\item{n}{Number of iteration}

\item{m}{mean value}

\item{s}{standard deviaiton}

\item{a}{Lower bound}

\item{b}{Upper bound}
}
\description{
Beta function
}
\author{
Jason Cope
}
