% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlnorm.R
\name{rtlnorm}
\alias{rtlnorm}
\title{Random lognormal}
\usage{
rtlnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{number of draws}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Random lognormal
}
\author{
Jason Cope
}
