% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_SS.R
\name{RUN.SS}
\alias{RUN.SS}
\title{Run SS function}
\usage{
RUN.SS(path, ss.exe = "ss", ss.cmd = " -nohess -nox")
}
\arguments{
\item{path}{location of the executable}

\item{ss.exe}{executable name for ss (e.g., "ss")}

\item{ss.cmd}{command line input used when running SS}
}
\description{
Run SS function
}
\author{
Jason Cope
}
