\name{split_multiple}
\alias{split_multiple}
\title{Splits a data...}
\usage{split_multiple(df, n, sep="_")}
\description{Splits a data.frame into multiple lines}
\details{When length is greater than 1, duplicates that row}
\keyword{internal}
\arguments{\item{df}{data.frame with .sss metadata}
\item{n}{Numeric vector that indicates the number of times each row must be repeated}
\item{sep}{character vector defining the string that separates field and subfield names}}
