% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-xml.R
\name{splitMultiple}
\alias{splitMultiple}
\title{Splits a data.frame into multiple lines}
\usage{
splitMultiple(df, n, sep = "_")
}
\arguments{
\item{df}{data.frame with .sss metadata}

\item{n}{Numeric vector that indicates the number of times each row must be repeated}

\item{sep}{character vector defining the string that separates field and subfield names}
}
\description{
When length is greater than 1, duplicates that row
}
\keyword{internal}
