\name{splitSSS}
\alias{splitSSS}
\title{Splits sss metadata into multiple lines}
\usage{
  splitSSS(sss, sep = "_")
}
\arguments{
  \item{sss}{data.frame containing .sss metadata}

  \item{sep}{character vector defining the string that
  separates field and subfield names}
}
\description{
  This is necessary when the variable type is "multiple"
}
\keyword{internal}

