% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSS_code_newSRs.r
\name{SSS}
\alias{SSS}
\title{This file contails all the SSS code}
\usage{
SSS(filepath, file.name, reps = 1000, seed.in = 19, Dep.in = c(1, 0.4,
  0.1), M.in = c(3, 0.1, 0.4, 3, 0.1, 0.4), SR_type = 3, h.in = c(1, 0.6,
  0.2), FMSY_M.in = c(1, 0.5, 0.1), BMSY_B0.in = c(1, 0.5, 0.1),
  L1.in = c(0, 0, 0, 0), Linf.in = c(0, 0, 0, 0), k.in = c(0, 0, 0, 0),
  Zfrac.Beta.in = c(-99, 0.2, 0.6, -99, 0.5, 2), R_start = c(0, 8),
  doR0.loop = c(1, 4.1, 12.1, 0.5), sum_age = 0, sb_ofl_yrs = c(2010,
  2011, 2012), f_yr = 2012, year0 = 1916, genders = F, BH_FMSY_comp = F)
}
\arguments{
\item{fmsy.m.in}{NOT YET FULLY IMPLEMENTED}

\item{bmsy.b0.in}{NOT YET FULLY IMPLEMENTED}
}
\description{
This file contails all the SSS code
}
\author{
Jason Cope and Chantel Wetzel
}
