% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-csv.R
\name{addQtext}
\alias{addQtext}
\title{Assigns question text to variable.labels attribute.}
\usage{
addQtext(sss, df)
}
\arguments{
\item{sss}{Parsed .sss metadata information}

\item{df}{Parsed .asc data}
}
\description{
Assigns question text to variable.labels attribute.
}
\keyword{internal}

