% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change_SSfiles.R
\name{Change.SSfiles}
\alias{Change.SSfiles}
\title{Change SS files}
\usage{
Change.SSfiles(main.dir, sppname.in, file.make = c(1, 1, 1, 1, 1),
  starter.par = "N")
}
\arguments{
\item{main.dir}{main directory}

\item{sppname.in}{species name}

\item{file.make}{vector that controls whether or not files are copied. Expected vector is c(par file, forecast file, data file, control file, ss executable). Value can be 1 = copy file. or 0 = do not copy}

\item{starter.par}{Y/N switch to use the par file}
}
\description{
Change SS files
}
\author{
Jason Cope
}
