% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-csv.R
\name{changeValues}
\alias{changeValues}
\title{Applies coded values to asc data, as described in sss metadata.}
\usage{
changeValues(sss, df)
}
\arguments{
\item{sss}{Parsed .sss metadata information}

\item{df}{Parsed .asc data}
}
\description{
Applies coded values to asc data, as described in sss metadata.
}
\keyword{Internal}
