% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-xml.R
\name{getSSSrecord}
\alias{getSSSrecord}
\title{Reads all "variables" inside the triple-s "record".}
\usage{
getSSSrecord(node)
}
\arguments{
\item{xmlNode}{XML node}
}
\description{
This function parses the record node, extracts all variables
and creates a data frame with information about size, position, type, etc.
}
\keyword{Internal}
