\name{namesMultiple}
\alias{namesMultiple}
\title{Reads all "codes" inside the triple-s "record"}
\usage{
  namesMultiple(names, length, sep = "_")
}
\arguments{
  \item{x}{XML node}
}
\description{
  This function parses the record node and extracts all
  "codes" nodes into a data.frame
}
\keyword{internal}

