\name{parse_sss_metadata}
\alias{parse_sss_metadata}
\title{Parses a triple-s XML (sss) metadata file, as specified by the triple-s XML standard.}
\usage{parse_sss_metadata(XMLdoc)}
\description{Parses a triple-s XML (sss) metadata file, as specified by the triple-s XML standard.}
\details{This function reads and parses a .sss XML metadata file as well as its associated .asc data file. The .sss standard defines a standard survey structure
#'}
\keyword{parse}
\seealso{read_sss_metadata, read_sss, read_sss_data}
\arguments{\item{XMLdoc}{An XML document - as returned by XML()}}
