\name{parse_sss_metadata}
\alias{parse_sss_metadata}
\title{Parses a triple-s XML (...}
\usage{parse_sss_metadata(XMLdoc)}
\description{Parses a triple-s XML (.sss) metadata file, as specified by the triple-s XML standard}
\details{This function parses a .sss XML metadata file.
The .sss standard defines a standard survey structure}
\keyword{XML}
\seealso{read_sss_metadata, read_sss, read_sss_data}
\arguments{\item{sss_filename}{Name of .sss file containing the survey metadata}}
\examples{# Not executed
# read_sss_metadata("sample.sss")}
