% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sss-metadata.R
\name{parseSSSmetadata}
\alias{parseSSSmetadata}
\title{Parses a triple-s XML (sss) metadata file, as specified by the triple-s XML standard.}
\usage{
parseSSSmetadata(x, XMLdoc)
}
\arguments{
\item{x}{An XML document - as returned by \code{\link[XML:xml]{XML::xml()}}, or \code{\link[=readSSSmetadata]{readSSSmetadata()}}}

\item{XMLdoc}{No longer used. Use \code{x} instead.}
}
\description{
This function reads and parses a .sss XML metadata file as well as its associated .asc data file. The .sss standard defines a standard survey structure
}
\seealso{
readSSSmetadata, read.sss, readSSSdata
}
\concept{parsing functions}
