\name{read_sss_data}
\alias{read_sss_data}
\title{Reads a triple-s XML (...}
\usage{read_sss_data(data_filename)}
\description{Reads a triple-s XML (.asc) data file, as specified by the triple-s XML standard}
\details{This function reads and parses a .asc XML data file.
The .sss standard defines a standard survey structure}
\keyword{XML}
\seealso{read_sss, read_sss_metadata}
\arguments{\item{asc_filename}{Name of .asc file containing the survey metadata}}
\examples{# Not executed
# read_sss_data("sample.asc")}
