% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sss-metadata.R
\name{readSSSdata}
\alias{readSSSdata}
\title{Reads a triple-s XML (asc) data file, as specified by the triple-s XML standard.}
\usage{
readSSSdata(x, ascFilename)
}
\arguments{
\item{x}{Name of .asc file containing the survey metadata}

\item{ascFilename}{No longer used. Use \code{x} instead.}
}
\description{
This function reads an `.asc`` data file.
}
\examples{
sampleRoot <- system.file("data/sample-1", package = "sss")
filenameSSS <- file.path(sampleRoot, "sample.sss")
filenameASC <- file.path(sampleRoot, "sample.asc")

readSSSdata(filenameSSS)
readSSSmetadata(filenameSSS)
}
\seealso{
\code{\link[=read.sss]{read.sss()}}, \code{\link[=readSSSmetadata]{readSSSmetadata()}}
}
\keyword{parse}
