\name{repN}
\alias{repN}
\title{Repeats each element n times}
\usage{
  repN(n)
}
\arguments{
  \item{n}{Numeric vector that indicates the number of
  times each row must be repeated}
}
\description{
  Each element is repeated n times.  This is used to
  construct a vector of the new length after accounting for
  fields of type multiple
}
\keyword{internal}

