\name{split_multiple}
\alias{split_multiple}
\title{Splits a data frame column into multiple columns...}
\usage{split_multiple(df_column, start, end)}
\description{Splits a data frame column into multiple columns}
\seealso{llply_colwise}
\keyword{internal}
\arguments{\item{df_column}{A fixed-width column in a data frame}
\item{start}{Numeric: start position}
\item{end}{Numeric: end position}}
