% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-xml.R
\name{splitSSS}
\alias{splitSSS}
\title{Splits sss metadata into multiple lines}
\usage{
splitSSS(sss, sep = "_")
}
\arguments{
\item{sss}{data.frame containing .sss metadata}

\item{sep}{character vector defining the string that separates field and subfield names}
}
\description{
This is necessary when the variable type is "multiple"
}
\keyword{internal}

