% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss-package.R
\docType{package}
\name{sss-package}
\alias{sss-package}
\alias{sss}
\title{Tools for importing files in the triple-s (Standard Survey Structure) format.}
\description{
sss is a set of tools to import survey files in the .sss (triple-s) format.  Triple-s is a standard to transfer survey data between applications.
}
\references{
http://www.triple-s.org/

The most important exported function is:
\itemize{
\item \code{\link[=read.sss]{read.sss()}}
}
}
\author{
Andrie de Vries
}
\keyword{package}
