% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss-package.R
\docType{package}
\name{sss-package}
\alias{sss}
\alias{sss-package}
\title{sss: Import Files in the Triple-s (Standard Survey Structure) Format}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to import survey files in the '.sss' (triple-s) format. The package provides the function 'read.sss()' that reads the '.asc' (or '.csv') and '.sss' files of a triple-s survey data file. See also \url{https://www.triple-s.org/}.
}
\references{
https://www.triple-s.org/

The most important exported function is:
\itemize{
\item \code{\link[=read.sss]{read.sss()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrie.github.io/sss/}
  \item Report bugs at \url{https://github.com/andrie/sss/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

}
\keyword{internal}
