\docType{package}
\name{sss-package}
\alias{"sss-package"}
\alias{sss}
\alias{sss-package}
\title{Tools for importing files in the triple-s .(Standard Survey Structure) format.}
\description{
  sss is a set of tools to import survey files in the .sss
  (triple-s) format.  triple-s is a standard to transfer
  survey data between applications.
}
\references{
  http://www.triple-s.org/

  The most important exported function is:
  \code{\link{read.sss}}
}
\keyword{package}

